+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  * Make the `TupleType` alias public
  * Add the `ValueType` enum that represents the type of a value for easier comparisons and matchings
  * Add `EvalexprResult<T>` type that uses the `EvalexprError` type (renamed from `Error`)
+ * Add `Node::eval_number` and `Node::eval_number_with_context` to evaluate to int or float and silently converting to float
 
 ### Removed